+++ R/targeting.h.R
@@ -195,7 +195,6 @@ targetingResults <- if (requireNamespace("jmvcore", quietly=TRUE)) R6::R6Class(
                 width=700,
                 height=700,
                 renderFun=".targetingPlot",
-                requiresData=TRUE,
                 clearWith=list(
                     "vars",
                     "panel",